cleanup/improvements
