--- XeCord/GTA5/Hooks.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "Invoker.h"
-#include <XexUtils.h>
-
-namespace Hooks {
-	XexUtils::Detour& GetMainDetour();
-	BOOL MainHook(Rage::Native::NativeContext* Context);
-}
