+++ XeCord/XeCord.vcxproj
@@ -224,6 +224,8 @@
       <PreprocessorDefinitions>_XBOX</PreprocessorDefinitions>
       <AdditionalIncludeDirectories>..\deps\XexUtils\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <UndefinePreprocessorDefinitions>NDEBUG</UndefinePreprocessorDefinitions>
+      <ObjectFileName>$(IntDir)%(RelativeDir)%(Filename)%(Extension).obj</ObjectFileName>
+      <AssemblerListingLocation>$(IntDir)%(RelativeDir)%(Filename).asm</AssemblerListingLocation>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
@@ -231,7 +233,7 @@
       <OptimizeReferences>true</OptimizeReferences>
       <ProgramDatabaseFile>$(OutDir)$(ProjectName).pdb</ProgramDatabaseFile>
       <SetChecksum>true</SetChecksum>
-      <AdditionalDependencies>xapilib.lib;d3d9.lib;d3dx9.lib;xgraphics.lib;xboxkrnl.lib;xnet.lib;xaudio2.lib;xact3.lib;x3daudio.lib;xmcore.lib;vcomp.lib</AdditionalDependencies>
+      <AdditionalDependencies>xapilib.lib;d3d9.lib;d3dx9.lib;xgraphics.lib;xboxkrnl.lib;xnet.lib;xonline.lib;xaudio2.lib;xact3.lib;x3daudio.lib;xmcore.lib;vcomp.lib</AdditionalDependencies>
       <AdditionalOptions>/DLL /ENTRY:"_DllMainCRTStartup" /ALIGN:128,4096 %(AdditionalOptions)</AdditionalOptions>
       <AdditionalLibraryDirectories>%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
     </Link>
@@ -258,6 +260,8 @@
       <PreprocessorDefinitions>_XBOX</PreprocessorDefinitions>
       <AdditionalIncludeDirectories>..\deps\XexUtils\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <UndefinePreprocessorDefinitions>NDEBUG</UndefinePreprocessorDefinitions>
+      <AssemblerListingLocation>$(IntDir)%(RelativeDir)%(Filename).asm</AssemblerListingLocation>
+      <ObjectFileName>$(IntDir)%(RelativeDir)%(Filename)%(Extension).obj</ObjectFileName>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
@@ -265,7 +269,7 @@
       <OptimizeReferences>true</OptimizeReferences>
       <ProgramDatabaseFile>$(OutDir)$(ProjectName).pdb</ProgramDatabaseFile>
       <SetChecksum>true</SetChecksum>
-      <AdditionalDependencies>xapilib.lib;d3d9.lib;d3dx9.lib;xgraphics.lib;xboxkrnl.lib;xnet.lib;xaudio2.lib;xact3.lib;x3daudio.lib;xmcore.lib;vcomp.lib</AdditionalDependencies>
+      <AdditionalDependencies>xapilib.lib;d3d9.lib;d3dx9.lib;xgraphics.lib;xboxkrnl.lib;xnet.lib;xonline.lib;xaudio2.lib;xact3.lib;x3daudio.lib;xmcore.lib;vcomp.lib</AdditionalDependencies>
       <AdditionalOptions>/DLL /ENTRY:"_DllMainCRTStartup" /ALIGN:128,4096 %(AdditionalOptions)</AdditionalOptions>
       <AdditionalLibraryDirectories>%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
     </Link>
@@ -302,9 +306,9 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="GameDB.cpp" />
-    <ClCompile Include="GTA5\Hooks.cpp" />
-    <ClCompile Include="GTA5\Invoker.cpp" />
-    <ClCompile Include="GTA5\SessionInfo.cpp" />
+    <ClCompile Include="GTA5_Hooks.cpp" />
+    <ClCompile Include="GTA5_Invoker.cpp" />
+    <ClCompile Include="SessionInfo.cpp" />
     <ClCompile Include="XeCord.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -313,9 +317,10 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="GTA5\Hooks.h" />
-    <ClInclude Include="GTA5\Invoker.h" />
-    <ClInclude Include="GTA5\SessionInfo.h" />
+    <ClInclude Include="Hooks.h" />
+    <ClInclude Include="Vector3.h" />
+    <ClInclude Include="GTA5_Invoker.h" />
+    <ClInclude Include="SessionInfo.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
